[REFACTOR/#169] pilgrimageCategoryQueryService 분리
변경 내역
- Category 관련 조회 서비스 분리
